refactor(api/products): 상수 파일 네이밍 수정
